# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "XXX系统后端 HTTP REST API",
        "title": "XXX SERVER API",
        "contact": {},
        "version": "0.0.1"
    },
    "basePath": "/xxx-server/v1",
    "paths": {
        "/setting": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "设置"
                ],
                "summary": "查询设置列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "设置所属的分栏，enum(预警-alert)",
                        "name": "section",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/resp.SettingsResponse"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "设置"
                ],
                "summary": "批量删除设置项",
                "parameters": [
                    {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "description": "ids",
                        "name": "ids",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    }
                }
            }
        },
        "/setting/alert/{adcode}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "设置"
                ],
                "summary": "查询特定地区预警设置",
                "parameters": [
                    {
                        "type": "string",
                        "description": "adcode",
                        "name": "adcode",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/resp.SettingResponse"
                        }
                    }
                }
            }
        },
        "/setting/shapefile": {
            "post": {
                "description": "使用multipart form格式上传zip：curl -X POST http://xxx -F \"file=@/home/test/test.zip\" -H \"Content-Type: multipart/form-data\"",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "设置"
                ],
                "summary": "将包含shp的zip压缩包转为WKT(srid=4326)",
                "parameters": [
                    {
                        "type": "file",
                        "description": "upload target",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    }
                }
            }
        },
        "/setting/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "设置"
                ],
                "summary": "获取单个设置项详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/resp.SettingResponse"
                        }
                    }
                }
            }
        },
        "/setting/{section}": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "设置"
                ],
                "summary": "新增设置项",
                "parameters": [
                    {
                        "type": "string",
                        "description": "设置所属的分栏",
                        "name": "section",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "设置的值",
                        "name": "value",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/resp.IdResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entity.Setting": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "ID",
                    "type": "integer"
                },
                "section": {
                    "description": "设置所属分栏，enum(预警-alert)",
                    "type": "string"
                },
                "value": {
                    "description": "设置的值",
                    "type": "object"
                }
            }
        },
...
        "resp.IdData": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "resp.IdResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "code",
                    "type": "integer",
                    "enum": [
                        200,
                        400,
                        401,
                        500
                    ]
                },
                "data": {
                    "$ref": "#/definitions/resp.IdData"
                },
                "msg": {
                    "description": "message",
                    "type": "string"
                }
            }
        },
        "resp.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "code",
                    "type": "integer",
                    "enum": [
                        200,
                        400,
                        401,
                        500
                    ]
                },
                "data": {
                    "description": "data",
                    "type": "object"
                },
                "msg": {
                    "description": "message",
                    "type": "string"
                }
            }
        },
        "resp.SettingResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "code",
                    "type": "integer",
                    "enum": [
                        200,
                        400,
                        401,
                        500
                    ]
                },
                "data": {
                    "$ref": "#/definitions/entity.Setting"
                },
                "msg": {
                    "description": "message",
                    "type": "string"
                }
            }
        },
        "resp.SettingsResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "code",
                    "type": "integer",
                    "enum": [
                        200,
                        400,
                        401,
                        500
                    ]
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Setting"
                    }
                },
                "msg": {
                    "description": "message",
                    "type": "string"
                }
            }
        }
    }
}